Use GetCurrentProcess to determine if CloseHandle required
On Windows 2012, we are seeing frequent errors during the finalizer for
Win32::Handle, which invokes CloseHandle:

'The handle is invalid' (error code 6)

Debugging and testing has shown that every instance of this error is
caused by the handle 18446744073709551615 -> "ffffffffffffffff".  We have also seen
that this matches the value passed into the constructor of Handle,
provided directly from GetCurrentHandle.

We were previously comparing this to the const "ffffffff" and skipping
the CloseHandle call on the assumption that "ffffffff" == (HANDLE)-1
based on the docs here: http://msdn.microsoft.com/en-us/library/windows/desktop/ms683179(v=vs.85).aspx

But given that we are directly seeing the GetCurrentHandle returning
"ffffffffffffffff", it seems that assumption was incorrect or at least has
changed.

 Instead of worrying about the correct value of that const, this PR
 makes it so that we're following the recommended approach in the link
 above - if we want to check if a handle == current process handle, we
 compare it to GetCurrentHandle instead of the const.

 In the process, it resolves the errors we're seeing on the Windows
 Server 2012 platform, where we are failing when CloseHandle returns
 "invalid handle".

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>
Signed-off-by: John McCrae <john.mccrae@progress.com>