Allow dsc_refresh_mode_disabled to be invoked in test
Currently we're failing this test on a pre-emptive direct execution of
Get-LocalDscConfigurationManager that we're doing in the test.  It is
not returning the expected result, but it's hiding the errors.

Instead, we'll let the actual thing we want to test occur
(dsc_refresh_mode_disabled) - because it runs the same command, and
raises an error if it fails.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>