zypper_package: Fix idempotency issue
Installing the apache2-mod_wsgi-python3 package is not idempotent due to the
fact that zypper will also return the apache2-mod_wsgi package. The
resolve_available_version method, incorrectly assumes this will only return one
item, but in this case it returns four items and will pick the first package
listed.

To work around this, we should verify that the package name listed in the output
matches what we are expecting to install.

Signed-off-by: Lance Albertson <lance@osuosl.org>