Enable state tracking on mutating Array operations
The State tracking feature is nice, especially as it allows to get
an event when an attribute is updated.

However, mutating operations on Arrays were not properly tracked, at
least for the "attribute_changed" event.
For instance, `default['array'] << 1` was not triggering an event.

The new code, ensure that all identified mutating methods are properly
implementing the State Tracking feature.
To do it, it leverages existing DISALLOWED_MUTATOR_METHODS constant.

Signed-off-by: Baptiste Courtois <b.courtois@criteo.com>