Move danger runner to dangerjs (#14134)
Danger Ruby just full on doesn't work with forked repos.  You can see
my and other's many attempts here:
https://github.com/danger/danger/issues/1103#issuecomment-1848222581

However, the JS version does in fact seem to work well. Move to that.

The syntax for the Dangerfile is a bit less pretty, and obviously for
a ruby project, we'd prefer a Ruby DSL, but working is better than
pretty.

Signed-off-by: Phil Dibowitz <phil@ipom.com>