Add `start_after_enable` property to `systemd_unit` resource
It is a common operational task is to start a service after enabling it. systemd
even has a flag for this (`--now`). I didn't see a way to cleanly add that flag
in the implementation so I added an explicit start.

Note this feature is different than the `:start` action which always starts the
service.

A `stop_after_disable` property may also be useful.