chocolatey_package: fix the use_choco_list logic
The logic for this was flipped in a refactor, and it would in fact use `choco
list` when not doing that was selected - it would only use the local disk
method if it was _not_ selected.  This does not cause a general functionality
regression, since it pushes it to the safest way, but it does cause a
significant (20s per chef run, in our environment) performance regression

This also fixes a corner case in that flow where a stray downcase could happen
on a nil, causing an exception.

Signed-off-by: Jaymz Julian <jaymzjulian@meta.com>