Avoid DeepMergeCache flap in loaded_recipe
Using self does a DeepMergeCache build on :recipes, which worsens as the array
grows with included recipes since appending the recipe to the array causes the
DeepMergeCache to drop for :recipes. loaded_recipe uses the :recipes array as a
an ordered set, and doing writes to :recipes outside of automatic precedence
doesn't make sense. Avoid reads from the other precedence levels entirely and
read directly from the automatic precedence.

Signed-off-by: David Crosby <dcrosby@fb.com>