Disable merging of chocolatey searches (#14327)
* Disable merging of chocolatey searches

There was a time when the following worked:

```
choco list -r firefox wireshark
choco search -r firefox wireshark
```

However, it turns out this only worked by accident, not by spec - and this has stopped working now.  It's known to not work, see https://github.com/chocolatey/choco/issues/2116 .  Additionally, the other potential workarounds are also not well supported.  So instead, we tighten the caching a little to at least avoid multiple API requests for the same package, and drop back to single package at a time searches for now, since that's what upstream chocolatey demands.  Once we can fix the issue in chocolatey itself, then we can bring this feature back.

(if you have just teh right server versions, which I did when I was testing, it does work - I was able to reproduce it as a regression, but it turns out what I considered "working" was, in fact, the bug and not the correct behaviour)

Note that this means now you see in debug:
```
[2024-04-10T00:36:54-07:00] INFO: Fetching chocolatey package list
[2024-04-10T00:36:54-07:00] DEBUG: Choco List Command: ["search", "-r", "firefox"]
[2024-04-10T00:36:55-07:00] DEBUG: Choco List Command: ["search", "-r", "wireshark"]
[2024-04-10T00:36:58-07:00] INFO: chocolatey_package[wireshark, firefox] upgraded(allow_downgrade) ["wireshark"] to ["3.0.1"]
```

which is correct per chocolatey spec, sadly.

Signed-off-by: Jaymz Julian <jaymzjulian@meta.com>

* Be consistent around taking about booleans in chocolatey_package

Co-authored-by: Thomas Powell <104777878+tpowell-progress@users.noreply.github.com>

Signed-off-by: Jaymz Julian <jaymzjulian@meta.com>

---------

Signed-off-by: Jaymz Julian <jaymzjulian@meta.com>
Co-authored-by: Jaymz Julian <jaymzjulian@meta.com>