[Immutablize#convert_value] return already frozen objects
One of the reasons DeepMergeCache flaps are so bad for performance is that we
dup new objects that aren't Hashes/Arrays during the creation of the cache, but
we also freeze the new objects well. Freezing an Object makes further
modification of the object result in RuntimeError, which means they are already
immutable.

Given the expectation that the objects aren't able to be further modified,
creating duplicate objects (which are then frozen) appears to be needless,
since we can instead pass the frozen object during ImmutableMash/ImmutableArray
creation.

References:
- https://docs.ruby-lang.org/en/master/Object.html#method-i-freeze
- https://docs.chef.io/release_notes_client/#13.0.113-freezing-immutable-merged-attributes

Signed-off-by: David Crosby <dcrosby@fb.com>