prefer extra_paths first when looking for bin file
When passing extra paths users would expect either command to be looked in the extra_path parameter first for cases of multiple bin commands existing on system or this should add tiny performance enhancement as likely the bin isn't in a default bin dir and user is adding extra dirs to lookin. Thus looking in them first will match early.