Include unextracted tarball instead of expanded files
In order to prevent warnings/errors and potential conflicts when we modify
or remove files that are managed by the package manager, this shows
using the unextracted tar.gz as the primary 'installed' file.

We expand it as part of 'postinst'; this prevents us from modifying the files that
are part of the package itself when we do uninstall and cleanup prior to
upgrade.

We're also specific about version and release throughout the spec now,
since it will be possible for more than one version to be installed
concurrently.  This ensures that the stubs generated for this package
point specifically to the versions included in this package.

This does not include the migration tool, which would be added as an
additional package file.

Important note: Increases the installed size over the base install
by the size of the taball (currently ~130MB)

Additional notes: I think there are some compatibility behaviors in
these scripts (orginally from omnibus) that we can let go of since we're
not supporting native installers for those platforms with this release.

Signed-off-by: Marc A. Paradise <marc.paradise@progress.com>