change default for quiet mode to false
this change will require users to explicitly set quiet mode to suppress output. since quiet mode was not working before the fix introduced in https://github.com/chef/chef/pull/14779, the idea here is to not surprise users not setting quiet mode explicitly by missing output from the logs when this fix is shipped. updating the quiet mode to false will ensure the users who are not setting the value continue seeing the same output as before.

Signed-off-by: rishichawda <rishichawda@users.noreply.github.com>