fix visual c++ depdendency issue for windows habitat package
the habitat package for windows was failing to run chef-client because of some missing dlls for chef powershell shim module. adding the package path to RUBY_DLL_PATH seems to fix this issue. this commit also removes the run hook for windows plan that was added previously to fix this issue in https://github.com/chef/chef/pull/14843 since that is not needed.

Signed-off-by: rishichawda <rishichawda@users.noreply.github.com>