disabling this value (#14800)
* disabling this value

this is a value that tells expeditor to do things with packages and omnibus builds

this is going to be one of many configuration updates before we get this ready to be promoted via expeditor

Signed-off-by: Sean Simmons <sean.simmons@progress.com>

* this is a major change to the project and how we promote

this will need to be discussed and well thought out. this is just a preliminary change to ensure we dont release an omnibus build with chef19.

Signed-off-by: Sean Simmons <sean.simmons@progress.com>

---------

Signed-off-by: Sean Simmons <sean.simmons@progress.com>