fix: multiple apt_repos with same key url should be signed-by (#15218)
Currently, when having multiple apt_repository resources that use the
same key url (e.g. Debian bookworm, bookworm-updates and
bookworm-backports), only the first apt_repository resource that gets
executed will have its key imported into a keyring.

The other apt_repository resources using the same key will not get
their keyrings imported, as the remote_file is up to date.

By notifying the import if the key does not exist, we work around this
problem.

I also removed the `:delete` notification of the keyring file when the
`remote_file` resource changes, as it caused unwanted deletes of the
actual keyring file _after_ the `gpg --import`.
This doesn't break anything, as `gpg --import` does not require the
keyring file to be missing, it just imports the new key in the same file
if it exists.

Signed-off-by: Frederik Thuysbaert <frederik.thuysbaert@combell.group>