[ChefRepositoryFileSystemCookbookEntry] Add children existence method (#15244)
This is kind of a "Facebook-scale" issue, but when you have lots of cookbooks and files, creating big arrays of child objects that are checked only for non-zero lengths can add whole seconds to Chef Zero cookbook resolution. Instead of creating the big arrays, simply detect if there's at least one child and return accordingly.

Signed-off-by: David Crosby <dcrosby@fb.com>