[package] Add DNF5 support to the dnf_package provider
This work is kindly sponsored by Meta Platforms, Inc.

This PR adds DNF5 support to the `package` (and `dnf_package`) provider.

There is significant overlap in dnf4 and dnf5 - while the backend is
entirely re-written from Python to C++, the Python bindings are
reasonably similar, and so it made sense to not make a new resource or a
new helper, and simply make the existing code handle both cases.

The majority of the work is in the python `dnf_helper.py` script. The
changes are:

- Remove the (never called, never worked) `flushcache()` method
- Add a debug mechanism via a env variable
- factor out get_sack to have a dnf4 and dnf5 backend
- rename query to query_dnf4 and add a query_dnf5
- Make style consistent

Other changes:

packages/dnf.rb
- determine version of dnf
- handle command line differences in adding, parsing, and removing
  version locks

package/dnf/python_helper.rb:
- Update test for "do we have the right libs" to handle dnf4 and dnf5

spec/functional/resource/dnf_package_spec.rb
- Fix bug in generated repo config (`enable` => `enabled`) - this is
  correct in both dnf4 and dnf5
- Determine version of dnf
- Update usage of `dnf config-manager` to handle differences in dnf5
- Add additional RPM macros needed in DNF5 to ignore our test RPMs
  which aren't signed
- Update to handle differences in version locks between 4 and 5

All unittests pass on both dnf4 and dnf5.

Signed-off-by: Phil Dibowitz <phil@ipom.com>